MFW-3413 Moved retrieve_data code to schema_utils
