+++ v1/policy_manager/validatepolicy.py
@@ -5,9 +5,8 @@ import json
 import jsonschema
 import os
 import unittest
-import pathlib
 import referencing
-from urllib.parse import urlsplit
+from v1.schema_utils.util import retrieve_data
 
 """
 TestPolicyManager tests all parts of the policy manager schema
@@ -64,25 +63,6 @@ class TestPolicyManager(unittest.TestCase):
             print("WARNING: Using default schema filename:" + str(schema_filename))
         with open(schema_filename, "r") as schema_fp:
             schema_data = json.load(schema_fp)
-
-        def retrieve_data(uri):
-            """
-            Retrieve data from a specified URI and return it as a referencing.Resource object.
-            This function processes the URI and retrieves data from it.
-            It handleslocal file ('file' scheme) URIs.
-
-            Args:
-            uri: The URI specifying the location of the data to retrieve.
-
-            Returns:
-            referencing.Resource: A Resource object representing the retrieved data.
-            """
-            parsed = urlsplit(uri)
-            if parsed.scheme == "file":
-                parsedpath = current_directory + parsed.path[1::]
-                path = pathlib.Path(parsedpath)
-            contents = json.loads(path.read_text())
-            return referencing.Resource.from_contents(contents)
 
         registry = referencing.Registry(retrieve=retrieve_data)
